# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SECOM MES Database API",
    "description": "REST API for SECOM Manufacturing Execution System Database\n\nThis API provides access to semiconductor manufacturing data including:\n- Master data (Equipment, Shifts, Operators, Product Types, Features)\n- Production data (Lots, Measurements, Quality Results)\n- Analytics (Production summaries, Equipment health, Shift performance, Quality analytics)\n\n**Database**: MariaDB 11.2\n**Total Records**: 1,567 lots with 590 sensor measurements each\n**Time Period**: July 2008 - September 2008\n\n**Key Endpoints**:\n- `/equipment` - Manufacturing equipment management\n- `/lots` - Production lot tracking with filters\n- `/measurements` - Sensor measurement data\n- `/quality` - Quality inspection results\n- `/analytics` - Analytics and reporting endpoints\n",
    "contact": {
      "name": "SECOM MES Team",
      "email": "mes@secom.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/products/{productTypeId}": {
      "get": {
        "tags": [
          "Product Types"
        ],
        "summary": "Get product type by ID",
        "description": "Returns a single product type by product type ID",
        "operationId": "getProductTypeById",
        "parameters": [
          {
            "name": "productTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product Types"
        ],
        "summary": "Update product type",
        "description": "Updates an existing product type record",
        "operationId": "updateProductType",
        "parameters": [
          {
            "name": "productTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product Types"
        ],
        "summary": "Delete product type",
        "description": "Deletes a product type record",
        "operationId": "deleteProductType",
        "parameters": [
          {
            "name": "productTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/operators/{operatorId}": {
      "get": {
        "tags": [
          "Operators"
        ],
        "summary": "Get operator by ID",
        "description": "Returns a single operator by operator ID",
        "operationId": "getOperatorById",
        "parameters": [
          {
            "name": "operatorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Operator"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Operators"
        ],
        "summary": "Update operator",
        "description": "Updates an existing operator record",
        "operationId": "updateOperator",
        "parameters": [
          {
            "name": "operatorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Operator"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Operator"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Operators"
        ],
        "summary": "Delete operator",
        "description": "Deletes an operator record",
        "operationId": "deleteOperator",
        "parameters": [
          {
            "name": "operatorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": [
          "Product Types"
        ],
        "summary": "Get all product types",
        "description": "Returns list of all product types",
        "operationId": "getAllProductTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product Types"
        ],
        "summary": "Create new product type",
        "description": "Creates a new product type record",
        "operationId": "createProductType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      }
    },
    "/operators": {
      "get": {
        "tags": [
          "Operators"
        ],
        "summary": "Get all operators",
        "description": "Returns list of all production operators",
        "operationId": "getAllOperators",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Operator"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Operators"
        ],
        "summary": "Create new operator",
        "description": "Creates a new operator record",
        "operationId": "createOperator",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Operator"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Operator"
                }
              }
            }
          }
        }
      }
    },
...
    "/products/family/{family}": {
      "get": {
        "tags": [
          "Product Types"
        ],
        "summary": "Get product types by family",
        "description": "Returns all product types in a specific family (Logic, Memory, Analog)",
        "operationId": "getProductTypesByFamily",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/code/{code}": {
      "get": {
        "tags": [
          "Product Types"
        ],
        "summary": "Get product type by code",
        "description": "Returns product type by unique product code",
        "operationId": "getProductTypeByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      }
    },
    "/operators/{operatorId}/lots": {
      "get": {
        "tags": [
          "Operators"
        ],
        "summary": "Get operator's lots",
        "description": "Returns all lots processed by a specific operator",
        "operationId": "getOperatorLots",
        "parameters": [
          {
            "name": "operatorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lot"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operators/department/{department}": {
      "get": {
        "tags": [
          "Operators"
        ],
        "summary": "Get operators by department",
        "description": "Returns all operators in a specific department",
        "operationId": "getOperatorsByDepartment",
        "parameters": [
          {
            "name": "department",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Operator"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operators/code/{code}": {
      "get": {
        "tags": [
          "Operators"
        ],
        "summary": "Get operator by code",
        "description": "Returns operator by unique operator code",
        "operationId": "getOperatorByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Operator"
                }
              }
            }
          }
        }
      }
    },
...
    "/equipment/type/{type}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get equipment by type",
        "description": "Returns all equipment of a specific type (CVD, Etcher, etc.)",
        "operationId": "getEquipmentByType",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equipment"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Shift": {
        "type": "object",
        "properties": {
          "shiftId": {
            "type": "integer",
            "format": "int32"
          },
          "shiftCode": {
            "type": "string"
          },
          "shiftName": {
            "type": "string"
          },
          "startTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "endTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Equipment": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "integer",
            "format": "int32"
          },
          "equipmentCode": {
            "type": "string"
          },
          "equipmentName": {
            "type": "string"
          },
          "equipmentType": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "installDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Lot": {
        "type": "object",
        "properties": {
          "lotId": {
            "type": "integer",
            "format": "int32"
          },
          "lotNumber": {
            "type": "string"
          },
          "productType": {
            "$ref": "#/components/schemas/ProductType"
          },
          "equipment": {
            "$ref": "#/components/schemas/Equipment"
          },
          "operator": {
            "$ref": "#/components/schemas/Operator"
          },
          "shift": {
            "$ref": "#/components/schemas/Shift"
          },
          "productionStart": {
            "type": "string",
            "format": "date-time"
          },
          "productionEnd": {
            "type": "string",
            "format": "date-time"
          },
          "waferCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Operator": {
        "type": "object",
        "properties": {
          "operatorId": {
            "type": "integer",
            "format": "int32"
          },
          "operatorCode": {
            "type": "string"
          },
          "operatorName": {
            "type": "string"
          },
          "employeeNumber": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "hireDate": {
            "type": "string",
            "format": "date"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProductType": {
        "type": "object",
        "properties": {
          "productTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "productCode": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "productFamily": {
            "type": "string"
          },
          "targetYield": {
            "type": "number"
          },
          "specificationVersion": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}